V4 regalloc: correct life-time hole handling.
When a life-range is active, but does not cover a position, it means
that it is inactive. When going from active to inactive (or vice-versa),
a (un-)spill is only needed when there is a stack-slot allocated to the
range. It is valid to not have a stack slot: when no call is done, or
when the register pressure does not push the end of the hole onto the
stack.

Change-Id: I5f5a05181c8cb384f031538dbefd7d11de682b8d
Reviewed-by: Lars Knoll <lars.knoll@digia.com>