V4: add counters for built-in function invocations.
When enabled, counters for the number of invocations of built-in runtime
functions are dumped when the application exits normally.

Note: this gives a penalty on execution speed. It is not meant to be
super light-weight, but as a tool to have an indication which functions
get called most, thereby helping to indicate which code-paths to
optimize.

Change-Id: Ica701f36a578affcce5ee8414532259972e5ede5
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>