Fix qqmllocale test to use correct spec and time zone
By including the +10:00 offset in the date strings the test is actually
creating Qt::OffsetFromUTC datetimes, not QT::LocalTime datetimes. The
only reason the test currently passes is due to a bug in the QLocale
datetime formatter which uses the current system daylight time
abbreviation in all cases regardless of the datetime spec or if
daylight or standard time.  Change 61774 fixes this bug but cannot
be merged as this test now fails.  This change fixes the bug by
removing the offset, causing Qt:LocalTime datetimes to be created
which pass the test both before and after change 61774.

See https://codereview.qt-project.org/#change,61774

Change-Id: I1642ab7c8aa516b703216f3791e2fc60866d1895
Reviewed-by: Alan Alpert <aalpert@blackberry.com>