Fix rendering errors in Qt Quick Controls gallery
First, when doing a partial rebuild, we need to mark
all elements for orphaning, not just those without a batch,
as the renderlist should at all times contain everything
that is rendered on screen.

Second, we had a small bug with material changes where we
would overwrite the rebuild state instead of adding to it.
This led to the render lists not being rebuilt from scratch
in some scenarios.

Third, after the shadow nodes rewrite, we didn't mark
for a full rebuild when clips were added.

Change-Id: I6707ebfa2908fcedd7fb636fb25b844be15fdea1
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@digia.com>