Disable the isPowerOfTwo functions that are only used with OpenGL ES 2
Found by ICC 14. This warning is not made an error because ICC is too
chatty when it comes to detecting unused functions. In this case,
however, it is right.

qsgtexturematerial.cpp(49): warning #177: function "isPowerOfTwo" was declared but never referenced
qsgtexture.cpp(84): warning #177: function "isPowerOfTwo" was declared but never referenced
qsgdefaultimagenode.cpp(368): warning #177: function "isPowerOfTwo" was declared but never referenced

Change-Id: I6c96dad3e1236fd33a02232f886ca33fa2b21258
Reviewed-by: Alan Alpert (Personal) <416365416c@gmail.com>
Reviewed-by: Gunnar Sletta <gunnar.sletta@digia.com>