[new compiler] Fix type name cache creation
Re-use the QQmlImports object created already in the type loader, which will
take care of importing scripts and others. This gets rid of the "###" in
qqmlobjectcreator and also delegates the type name cache population back
to the loader thread. (And will allow for passing to the JS code gen, for
type resolution and propagation in the future)

Change-Id: I21b0dfc853166200b3bc864f6c0644b2f87a4586
Reviewed-by: Lars Knoll <lars.knoll@digia.com>