Fix passing of exception table pointers to ARM runtime on unwinding
Our synthetic exception unwind table for ARM is located at

    (char *)codeStart + function->codeSize;

This relies on function->codeSize to contain the number of bytes of
instructions the function has, not the size of the MacroAssemblerCodeRef
(which contains the size of the entire area).

This patch fixes the calculation of function->codeSize and also replaces
the QHash for the IR::Function* -> CodeRef mapping in the masm backend
with a simple vector that's perfectly sufficient.

Bug spotted by Petr Nejedly

Change-Id: I78a53599085c613c6d97aa2490922f54e0bb4f63
Reviewed-by: Lars Knoll <lars.knoll@digia.com>