Convert ReturnedValue into a primitive (typedef to quint64)
ReturnedValue is used to return values from runtime methods
The type has to be a primitive type (no struct or union), so that
the compiler will return it in a register on all platforms.

They will be returned in rax on x64, [eax,edx] on x86 and [r0,r1] on arm.

Change-Id: I38433e6fad252370dda5dc335d9c5be8f22e8c76
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>