[new compiler] Fix function and binding expression setup
Binding expressions and QML used to be set up so that they were written
as function closure:

    (function(...) {
      expression here
    })

and then evaluated inside qml scope. With the new setup we do that closure
setup manually now.

For that we have to define a dummy outter "context scope" function in the
codegen, that will later be used to look up the context ids.

Change-Id: I0656419d67a1728451fcd46f402b03979f118b0b
Reviewed-by: Lars Knoll <lars.knoll@digia.com>