A dynamically created Window can have a parent Window
So far the parent relationship has existed only for Items.  parent
is still not exposed as a property of Window, but since it was possible
to give a parent parameter to Component.createObject(), it makes
sense to try to interpret it as a Window in that case.  So now a
Window can be created with another Window as its parent just as an
Item can be created with a parent Item.

Task-number: QTBUG-33644
Change-Id: I796198a38bd47253eef462c80f5098825451c59c
Reviewed-by: Alan Alpert (Personal) <416365416c@gmail.com>