Make GridView/ListView attached "view" property available to all delegate types
This is useful for accessing the view (e.g. "width", "cellWidth", ... properties)
from within section (ListView only), header, footer and highlight delegate components.

A typical usecase are components that are used in multiple views and therefore
cannot use the views's id for access.

The only attached property valid for those non-item delegates is "view". This
has been added to the documentation.

Change-Id: I33d976da778be23ed531a3b193ceee95ed9800d2
Task-number: QTBUG-32836
Reviewed-by: Alan Alpert <aalpert@blackberry.com>