Doc: Fixed broken links
There were many links (including the broken) on the Qt Quick
index page and QML type reference page using the html file name
to link to a page or a section within a page. Using html file
reference in a link is not a good approach as qdoc does not give
complete control to the author to define the html file name. This
change updates all such links to use the page \title and \target
names, which are considered to be better ways to link.

Task-number:QTBUG-33360
Change-Id: Ibd1b1e3223a7e205bbdcf21b1b7f6e515749c5c9
Reviewed-by: Nico Vertriest <nico.vertriest@digia.com>
Reviewed-by: Topi Reiniö <topi.reinio@digia.com>
Reviewed-by: Martin Smith <martin.smith@digia.com>
Reviewed-by: Jerome Pasion <jerome.pasion@digia.com>