QtQuick.Dialogs: fake window decoration clips, is opaque for touch
This is the wrapper for dialogs on platforms that don't have top-level
windows.  Mouse events over any unused area in the dialog would fall
through to the surrounding mouse area and close the dialog.  Now they
don't.  Clipping is enabled to ensure that dialog contents don't
overflow the "window" frame.

Change-Id: Ia3c0ae44bb47d46981daac9febd4cc93228dd20f
Reviewed-by: Liang Qi <liang.qi@digia.com>